requirements.txt
easyocr==1.6.2
Flask==2.2.3
matplotlib==3.7.1
mysql_connector_repackaged==0.3.1
numpy==1.24.2
opencv_contrib_python==4.5.4.60
opencv_python==4.10.0.84
pandas==2.0.0
Pillow==9.4.0
Pillow==10.3.0
scipy==1.9.1
seaborn==0.13.2
tensorflow==2.15.0
tensorflow_intel==2.15.0
